refactor(Frontend): Move distributed query management logic into nes-distributed
