+++ nes-distributed/CMakeLists.txt
@@ -1,0 +1,26 @@
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#    https://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+add_subdirectory(src)
+get_source(nes-distributed NES_DISTRIBUTED_SOURCE_FILES)
+
+add_library(nes-distributed ${NES_DISTRIBUTED_SOURCE_FILES})
+target_link_libraries(nes-distributed PUBLIC nes-common nes-single-node-worker-interface nes-logical-operators)
+target_include_directories(nes-distributed
+        PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
+
+if (NES_ENABLE_PRECOMPILED_HEADERS)
+    target_precompile_headers(nes-distributed REUSE_FROM nes-common)
+    target_compile_options(nes-distributed PUBLIC "-fPIC")
+endif ()
+
+add_tests_if_enabled(tests)
