+++ nes-distributed/include/WorkerCatalog.hpp
