+++ nes-distributed/include/WorkerConfig.hpp
