+++ nes-distributed/src/DistributedQueryIdGenerator.cpp
