+++ nes-distributed/src/NetworkTopology.cpp
