+++ nes-distributed/src/WorkerCatalog.cpp
