+++ nes-runtime/include/Listeners/AbstractQueryStatusListener.hpp
@@ -15,7 +15,7 @@
 
 #include <chrono>
 #include <Identifiers/Identifiers.hpp>
-#include <Runtime/Execution/QueryStatus.hpp>
+#include <QueryStatus.hpp>
 #include <Runtime/QueryTerminationType.hpp>
 #include <ErrorHandling.hpp>
 #include <QueryId.hpp>
