+++ nes-single-node-worker/interface/QueryStatus.hpp
@@ -1,0 +1,60 @@
+/*
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+#pragma once
+
+#include <chrono>
+#include <cstdint>
+#include <optional>
+#include <ostream>
+#include <Identifiers/Identifiers.hpp>
+#include <Util/Logger/Formatter.hpp>
+#include <magic_enum/magic_enum.hpp>
+#include <ErrorHandling.hpp>
+#include <QueryId.hpp>
+
+namespace NES
+{
+enum class QueryState : uint8_t
+{
+    Registered,
+    Started,
+    Running, /// Deployed->Running when calling start()
+    Stopped, /// Running->Stopped when calling stop() and in Running state
+    Failed,
+};
+
+inline std::ostream& operator<<(std::ostream& ostream, const QueryState& status)
+{
+    return ostream << magic_enum::enum_name(status);
+}
+
+struct QueryMetrics
+{
+    std::optional<std::chrono::system_clock::time_point> start;
+    std::optional<std::chrono::system_clock::time_point> running;
+    std::optional<std::chrono::system_clock::time_point> stop;
+    std::optional<Exception> error;
+};
+
+/// Summary structure of the query log for a query
+struct LocalQueryStatus
+{
+    QueryId queryId = INVALID_QUERY_ID;
+    QueryState state = QueryState::Registered;
+    QueryMetrics metrics{};
+};
+}
+
+FMT_OSTREAM(NES::QueryState);
